Fixed warnings and C++ improvements.
Fixed warnings in x64 as well as x86 on visual studio platforms.
Some improvements to the code to make use of C++11 features. (e.g. std::make_unique, = default/delete for constructors and destructors, override, etc.)

+++ RadeonRays/src/intersector/intersector_bittrail.cpp
@@ -340,7 +340,7 @@ namespace RadeonRays
                 // Besides that we need to permute the faces accorningly to BVH reordering, whihc
                 // is contained within bvh.primids_
                 int const* reordering = m_bvh->GetIndices();
-                for (int i = 0; i < numindices; ++i)
+                for (size_t i = 0; i < numindices; ++i)
                 {
                     int indextolook4 = reordering[i];
